+++ lib/src/postgrest_rpc_builder.dart
@@ -6,7 +6,7 @@ class PostgrestRpcBuilder extends PostgrestBuilder {
     String url, {
     Map<String, String>? headers,
     String? schema,
-    BaseClient? httpClient,
+    Client? httpClient,
   }) : super(
           url: Uri.parse(url),
           headers: headers ?? {},